Refactoring, mostly
