+++ microprofile/metrics/src/main/java/io/helidon/microprofile/metrics/MetricsCdiExtension.java
@@ -51,8 +51,10 @@ import io.helidon.metrics.api.RegistryFactory;
 import io.helidon.metrics.serviceapi.MetricsSupport;
 import io.helidon.microprofile.metrics.MetricAnnotationInfo.RegistrationPrep;
 import io.helidon.microprofile.metrics.MetricUtil.LookupResult;
-import io.helidon.microprofile.metrics.api.MetricAnnotationDiscoveryObserverProvider;
-import io.helidon.microprofile.metrics.api.MetricRegistrationObserverProvider;
+import io.helidon.microprofile.metrics.spi.MetricAnnotationDiscoveryObserver;
+import io.helidon.microprofile.metrics.spi.MetricAnnotationDiscoveryObserverProvider;
+import io.helidon.microprofile.metrics.spi.MetricRegistrationObserver;
+import io.helidon.microprofile.metrics.spi.MetricRegistrationObserverProvider;
 import io.helidon.microprofile.server.ServerCdiExtension;
 import io.helidon.servicecommon.restcdi.HelidonRestCdiExtension;
 import io.helidon.webserver.Routing;
