+++ microprofile/metrics/src/main/java/io/helidon/microprofile/metrics/spi/MetricRegistrationObserver.java
@@ -12,11 +12,10 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- *
  */
-package io.helidon.microprofile.metrics;
+package io.helidon.microprofile.metrics.spi;
 
-import io.helidon.microprofile.metrics.MetricAnnotationDiscoveryObserver.MetricAnnotationDiscovery;
+import io.helidon.microprofile.metrics.MetricAnnotationDiscovery;
 
 import org.eclipse.microprofile.metrics.Metadata;
 import org.eclipse.microprofile.metrics.Metric;
@@ -36,7 +35,7 @@ public interface MetricRegistrationObserver {
     /**
      * Notifies the observer that a metric has been registered due to a metric annotation that applies to an executable.
      *
-     * @param discovery the {@link MetricAnnotationDiscovery}
+     * @param discovery the {@link io.helidon.microprofile.metrics.MetricAnnotationDiscovery}
      *                  triggering this metric registration
      * @param metadata the metrics {@link org.eclipse.microprofile.metrics.Metadata} for the indicated metric
      * @param metricId the {@link org.eclipse.microprofile.metrics.MetricID} for the indicated metric
