+++ microprofile/metrics/src/test/resources/META-INF/services/io.helidon.microprofile.metrics.spi.MetricRegistrationObserverProvider
